# internal/rangekey: optimize Coalesce 