--- a/sstable/testdata/writer_range_keys
+++ b/sstable/testdata/writer_range_keys
@@ -44,7 +44,7 @@ SET a-c @2=foo
 SET a-c @1=bar
 SET a-c @3=baz
 ----
-a-c:{(#0,RANGEKEYSET,@3,baz) (#0,RANGEKEYSET,@2,foo) (#0,RANGEKEYSET,@1,bar)}
+a-c:{(#0,RANGEKEYSET,@2,foo) (#0,RANGEKEYSET,@1,bar) (#0,RANGEKEYSET,@3,baz)}
 
 # Aligned spans, mixed range key kinds.
 #
@@ -81,9 +81,9 @@ SET f-i @9=v9
 DEL f-i
 ----
 a-c:{(#0,RANGEKEYSET,@1,v1) (#0,RANGEKEYUNSET,@5)}
-c-d:{(#0,RANGEKEYSET,@5,v5) (#0,RANGEKEYSET,@2,v2)}
+c-d:{(#0,RANGEKEYSET,@2,v2) (#0,RANGEKEYSET,@5,v5)}
 d-e:{(#0,RANGEKEYSET,@9,v9) (#0,RANGEKEYDEL)}
-f-i:{(#0,RANGEKEYSET,@9,v9) (#0,RANGEKEYSET,@8,v8) (#0,RANGEKEYSET,@7,v7) (#0,RANGEKEYSET,@5,v5) (#0,RANGEKEYSET,@4,v4) (#0,RANGEKEYUNSET,@6) (#0,RANGEKEYDEL)}
+f-i:{(#0,RANGEKEYSET,@4,v4) (#0,RANGEKEYSET,@5,v5) (#0,RANGEKEYSET,@7,v7) (#0,RANGEKEYSET,@8,v8) (#0,RANGEKEYSET,@9,v9) (#0,RANGEKEYUNSET,@6) (#0,RANGEKEYDEL)}
 
 # Merge overlapping RANGEKEYSETs.
 #
@@ -100,7 +100,7 @@ SET b-f @2=bar
 SET c-d @3=baz
 ----
 a-b:{(#0,RANGEKEYSET,@1,foo)}
-b-c:{(#0,RANGEKEYSET,@2,bar) (#0,RANGEKEYSET,@1,foo)}
+b-c:{(#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYSET,@2,bar)}
 c-d:{(#0,RANGEKEYSET,@3,baz) (#0,RANGEKEYSET,@2,bar)}
 d-f:{(#0,RANGEKEYSET,@2,bar)}
 
@@ -108,8 +108,8 @@ d-f:{(#0,RANGEKEYSET,@2,bar)}
 #
 #  ^
 #  |                      •―――――○     [l,o) DEL
-#  |        •―――――――――――――――――――――――○ [e,q) SET   @4=baz
-#  |    •―○                           [c,d) DEL
+#  |        •―――――――――――――○           [e,l) SET   @4=baz
+#  |                              •―○ [p,q) DEL
 #  |    •―――――――――○                   [c,h) UNSET @3
 #  |  •―○                             [b,c) SET   @2=bar
 #  |•―――――――――――――――○                 [a,i) SET   @1=foo
@@ -125,41 +125,11 @@ SET e-q @4=baz
 DEL l-o
 ----
 a-b:{(#0,RANGEKEYSET,@1,foo)}
-b-c:{(#0,RANGEKEYSET,@2,bar) (#0,RANGEKEYSET,@1,foo)}
+b-c:{(#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYSET,@2,bar)}
 c-d:{(#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYUNSET,@3) (#0,RANGEKEYDEL)}
 d-e:{(#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYUNSET,@3)}
-e-h:{(#0,RANGEKEYSET,@4,baz) (#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYUNSET,@3)}
-h-i:{(#0,RANGEKEYSET,@4,baz) (#0,RANGEKEYSET,@1,foo)}
+e-h:{(#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYSET,@4,baz) (#0,RANGEKEYUNSET,@3)}
+h-i:{(#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYSET,@4,baz)}
 i-l:{(#0,RANGEKEYSET,@4,baz)}
 l-o:{(#0,RANGEKEYSET,@4,baz) (#0,RANGEKEYDEL)}
 o-q:{(#0,RANGEKEYSET,@4,baz)}
-
-build
-UNSET a-b @1
-SET a-b @1=foo
-----
-a-b:{(#0,RANGEKEYSET,@1,foo)}
-
-# Within the same fragmented span, key kinds do not affect one another. SETs and
-# DELs will have their own records. A SET will eliminate an UNSET as the former
-# sorts before the latter in a fragmented and coalesced span with the same
-# sequence number.
-#
-#  ^
-#  |        •―――――――○        [c,e) SET   @2=bar
-#  |    •―――――――○            [b,d) UNSET @2
-#  |    •―――――――○            [b,d) UNSET @1
-#  |•―――――――○                [a,c) SET   @1=foo
-#  |___________________________________
-#   a   b   c   d   e   f   g   h   i
-
-build
-SET a-c @1=foo
-UNSET b-d @1
-UNSET b-d @2
-SET c-e @2=bar
-----
-a-b:{(#0,RANGEKEYSET,@1,foo)}
-b-c:{(#0,RANGEKEYSET,@1,foo) (#0,RANGEKEYUNSET,@2)}
-c-d:{(#0,RANGEKEYSET,@2,bar) (#0,RANGEKEYUNSET,@1)}
-d-e:{(#0,RANGEKEYSET,@2,bar)}