Optimize rangekey.Coalesce by sorting range keys altogether once, and then
deduplicating in a second pass.

```
name                                   old time/op  new time/op  delta
Transform/shadowing=false/keys=1-10    46.5ns ± 0%  35.7ns ± 0%  -23.19%  (p=0.016 n=5+4)
Transform/shadowing=false/keys=2-10     116ns ± 0%    96ns ± 0%  -17.50%  (p=0.008 n=5+5)
Transform/shadowing=false/keys=4-10     436ns ± 0%   256ns ± 0%  -41.30%  (p=0.008 n=5+5)
Transform/shadowing=false/keys=8-10    1.77µs ± 0%  0.95µs ± 0%  -46.67%  (p=0.008 n=5+5)
Transform/shadowing=false/keys=16-10   8.85µs ± 0%  3.24µs ± 0%  -63.42%  (p=0.008 n=5+5)
Transform/shadowing=false/keys=32-10   48.7µs ± 0%   7.2µs ± 0%  -85.19%  (p=0.008 n=5+5)
Transform/shadowing=false/keys=64-10    170µs ± 0%    19µs ± 0%  -88.85%  (p=0.008 n=5+5)
Transform/shadowing=false/keys=128-10   432µs ± 0%    47µs ± 0%  -89.05%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=1-10     46.6ns ± 0%  36.0ns ± 2%  -22.83%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=2-10     95.6ns ± 1%  85.3ns ± 0%  -10.79%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=4-10      227ns ± 0%   201ns ± 0%  -11.68%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=8-10      774ns ± 0%   830ns ± 0%   +7.34%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=16-10    3.29µs ± 0%  2.79µs ± 0%  -15.23%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=32-10    16.3µs ± 0%   6.7µs ± 0%  -58.72%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=64-10    83.5µs ± 0%  18.9µs ± 0%  -77.42%  (p=0.008 n=5+5)
Transform/shadowing=true/keys=128-10    243µs ± 0%    48µs ± 0%  -80.14%  (p=0.008 n=5+5)
```

Close #2175.